clean up storybook * add modal decorator * add navigation decorator
