+++ storybook/decorators/index.js
@@ -1,0 +1,5 @@
+// @flow
+
+export {withModalProps} from './withModalProps'
+export {withProvider} from './withProvider'
+export {withNavigationProps} from './withNavigationProps'
