+++ storybook/decorators/withProvider.js
@@ -1,8 +1,9 @@
 // @flow
+
 import React from 'react'
 import {Provider} from 'react-redux'
 
-import configureStore from '../src/helpers/configureStore.js'
+import configureStore from '../../src/helpers/configureStore.js'
 
 import type {Node} from 'react'
